J_Search/requirements.txt
mysql-connector-python
streamlit
Flask